fiddling with pytransit
