+++ .ci/travis.sh
@@ -15,19 +15,14 @@ fi
 # Display some info
 conda info -a
 
-# Install some dependencies
+# Install batman
 pip install batman-package
 
-# Install pytransit and run all its scripts
-# now. PyTransit occasionally segfaults, so let's
-# isolate that from the rest of the build.
+# Install pytransit
 pushd $HOME
 git clone https://github.com/hpparvi/pytransit.git
 cd pytransit
 python setup.py config_fc --fcompiler=gnu95 --opt="-Ofast" --f90flags="-cpp -fopenmp -march=native" build install
-popd
-pushd $TRAVIS_BUILD_DIR/tex/figures/python/pytransit
-python compare_to_pytransit.py
 popd
 
 # Attempt to resolve issues with SSL certificate expiring for purl.org:
