--- tex/figures/python/pytransit/.gitignore
@@ -1,1 +1,0 @@
-*.txt
