--- tex/figures/python/pytransit_hack
@@ -1,12 +1,0 @@
-import time
-import pytransit
-import numpy as np
-u_g = np.loadtxt("u_g.txt")
-b = np.loadtxt("b.txt")
-number = 30
-m = pytransit.Gimenez(nldc=len(u_g), interpolate=False, nthr=0)
-tstart = time.time()
-for k in range(number):
-    pytransit_flux = m(b, 0.1, u_g)
-print((time.time() - tstart) / number)
-np.savetxt("flux_pytransit.txt", X=pytransit_flux)
